+++ app/src/main/AndroidManifest.xml
@@ -1,17 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
-    <uses-sdk android:targetSdkVersion="33" android:maxSdkVersion="28" />
 
+    <uses-sdk
+        android:maxSdkVersion="28"
+        android:targetSdkVersion="33" />
 
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:enableOnBackInvokedCallback="true"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
+        android:enableOnBackInvokedCallback="true"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -20,7 +22,10 @@
         android:theme="@style/Theme.Team99"
         tools:targetApi="31">
         <activity
-            android:name=".VideoDetailActivity"
+            android:name=".SplashScreenActivity"
+            android:exported="false" />
+        <activity
+            android:name=".Detail.VideoDetailActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
