--- app/src/main/java/com/example/team99/Constants.kt
@@ -1,8 +1,0 @@
-package com.example.team99
-
-object Constants {
-
-    var BASE_URL = "https://youtube.googleapis.com/"
-
-    var API_KEY = ">AIzaSyBx5x3nhrglEpE6nZqj37ywin9WJW9WhDc"
-}
