+++ app/src/main/java/com/example/team99/DTO/YoutubeVideosApi.kt
@@ -68,14 +68,6 @@ data class YoutubeVideosApi(
             data class Thumbnails(
                 @SerializedName("default")
                 val default: Default?,
-                @SerializedName("high")
-                val high: High?,
-                @SerializedName("maxres")
-                val maxres: Maxres?,
-                @SerializedName("medium")
-                val medium: Medium?,
-                @SerializedName("standard")
-                val standard: Standard?
             ) {
                 data class Default(
                     @SerializedName("height")
@@ -86,41 +78,6 @@ data class YoutubeVideosApi(
                     val width: Int?
                 )
 
-                data class High(
-                    @SerializedName("height")
-                    val height: Int?,
-                    @SerializedName("url")
-                    val url: String?,
-                    @SerializedName("width")
-                    val width: Int?
-                )
-
-                data class Maxres(
-                    @SerializedName("height")
-                    val height: Int?,
-                    @SerializedName("url")
-                    val url: String?,
-                    @SerializedName("width")
-                    val width: Int?
-                )
-
-                data class Medium(
-                    @SerializedName("height")
-                    val height: Int?,
-                    @SerializedName("url")
-                    val url: String?,
-                    @SerializedName("width")
-                    val width: Int?
-                )
-
-                data class Standard(
-                    @SerializedName("height")
-                    val height: Int?,
-                    @SerializedName("url")
-                    val url: String?,
-                    @SerializedName("width")
-                    val width: Int?
-                )
             }
         }
     }
