+++ app/src/main/java/com/example/team99/Home/HomeAdapter/CategoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.team99.Home
+package com.example.team99.Home.HomeAdapter
 
 import android.content.Context
 import android.content.Intent
@@ -10,7 +10,8 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.team99.MainActivity
-import com.example.team99.VideoDetailActivity
+import com.example.team99.Detail.VideoDetailActivity
+import com.example.team99.Home.Items.VideoItem
 import com.example.team99.databinding.VideoItemBinding
 
 class CategoryAdapter (private val mContext: Context) : RecyclerView.Adapter<CategoryAdapter.CategoryHolder>() {
