+++ app/src/main/java/com/example/team99/Home/HomeAdapter/ChannelAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.team99.Home
+package com.example.team99.Home.HomeAdapter
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -7,6 +7,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.team99.Home.Items.ChannelItem
 import com.example.team99.databinding.VideoItemBinding
 
 class ChannelAdapter (private val mContext: Context) : RecyclerView.Adapter<ChannelAdapter.ChanelHolder>() {
