+++ app/src/main/java/com/example/team99/Home/Items/VideoItem.kt
@@ -1,4 +1,4 @@
-package com.example.team99.Home
+package com.example.team99.Home.Items
 
 
 import java.io.Serializable
