+++ app/src/main/java/com/example/team99/Retrofit/Retrofit_interface.kt
@@ -1,8 +1,8 @@
 package com.example.team99.Retrofit
 
-import com.example.team99.YoutubeCategoriesApi
-import com.example.team99.YoutubeChannelApi
-import com.example.team99.YoutubeVideosApi
+import com.example.team99.DTO.YoutubeCategoriesApi
+import com.example.team99.DTO.YoutubeChannelApi
+import com.example.team99.DTO.YoutubeVideosApi
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
