+++ app/src/main/res/layout/myvd_video.xml
@@ -26,13 +26,14 @@
 
     <TextView
         android:id="@+id/myvd_video_title"
-        android:layout_width="wrap_content"
+        android:layout_width="180dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="18dp"
         android:layout_marginTop="10dp"
         android:textColor="@color/white"
         android:maxLines="2"
+        android:ellipsize="end"
         android:text="동영상 제목"
-        android:layout_marginStart="18dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/myvd_video_thumbnail" />
