+++ app/src/main/res/layout/video_item.xml
@@ -6,28 +6,39 @@
     android:layout_width="wrap_content"
     android:layout_height="wrap_content">
 
-
+    <androidx.cardview.widget.CardView
+        android:id="@+id/cardview_item"
+        android:layout_width="200dp"
+        android:layout_height="148dp"
+        app:cardCornerRadius="10dp"
+        app:cardElevation="0dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        >
     <ImageView
         android:id="@+id/thumbnails_iv"
-        android:layout_width="200dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
         android:adjustViewBounds="true"
+        android:scaleType="fitXY"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         tools:srcCompat="@tools:sample/avatars" />
 
+</androidx.cardview.widget.CardView>
+
     <TextView
         android:id="@+id/title_tv"
-        android:layout_width="wrap_content"
+        android:layout_width="200dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:layout_marginStart="8dp"
+        android:layout_marginStart="4dp"
+        android:layout_marginEnd="4dp"
         android:text="title"
         android:textColor="@color/white"
-        android:ems="12"
         android:maxLines="2"
         android:ellipsize="end"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/thumbnails_iv" />
+        app:layout_constraintTop_toBottomOf="@+id/cardview_item" />
 </androidx.constraintlayout.widget.ConstraintLayout>
